# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: AI Gateway API
  description: |
    Enterprise AI Gateway — 应用层 AI 治理中间件。
    提供 OpenAI 兼容的推理接口和多维度管理 API。

    ## 快速入门
    ```bash
    export GATEWAY_URL=http://localhost:8080
    export API_KEY=sk-your-key

    # 健康检查
    curl $GATEWAY_URL/admin/health

    # Chat Completion
    curl $GATEWAY_URL/v1/chat/completions \
      -H "Authorization: Bearer $API_KEY" \
      -d '{"model":"gpt-4","messages":[{"role":"user","content":"Hello"}]}'
    ```
  version: "2.1.0"
  contact:
    name: AI Gateway Team
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://gateway.example.com
    description: Production
paths:
...
  /v1/chat/completions:
    post:
      summary: Chat Completion (OpenAI compatible)
      description: |
        Streams or returns a chat completion from the routed upstream provider.
        Supports streaming via SSE when `stream: true`.
      operationId: createChatCompletion
      security:
        - ApiKeyAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
            example: Bearer sk-your-key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model:
                  type: string
                  example: gpt-4
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                        enum: [system, user, assistant]
                      content:
                        type: string
                  example:
                    - role: user
                      content: Hello
                stream:
                  type: boolean
                  default: false
                temperature:
                  type: number
                  format: float
                  default: 1.0
                max_tokens:
                  type: integer
                top_p:
                  type: number
                  format: float
                stop:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Successful response (stream or non-stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/embeddings:
    post:
      summary: Create Embeddings (OpenAI compatible)
      operationId: createEmbedding
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, input]
              properties:
                model:
                  type: string
                input:
                  oneOf:
                    - type: string
                    - type: array
                      items:
                        type: string
      responses:
        "200":
          description: Embedding response

  # ============================================
  # Management API — System
  # ============================================
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: API key from Authorization header
  schemas:
...
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
                  code:
                    type: string
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                  type:
                    type: string
...